@@ -17,108 +17,113 @@
       "script:generateGoogleActionsTypes": "ts-node scripts/generateGoogleActionsTypes.ts src/common/GoogleActionsTypes.ts"
    },
    "devDependencies": {
-      "@babel/core": "^7.7.5",
-      "@babel/preset-env": "^7.7.6",
-      "@graphql-codegen/cli": "^1.11.2",
-      "@graphql-codegen/fragment-matcher": "^1.15.0",
-      "@graphql-codegen/typescript": "^1.1.1",
-      "@graphql-codegen/typescript-operations": "^1.1.1",
-      "@graphql-codegen/typescript-resolvers": "^1.1.1",
+      "@babel/core": "^7.10.2",
+      "@babel/preset-env": "^7.10.2",
+      "@graphql-codegen/cli": "^1.15.1",
+      "@graphql-codegen/fragment-matcher": "^1.15.1",
+      "@graphql-codegen/typescript": "^1.15.1",
+      "@graphql-codegen/typescript-operations": "^1.15.1",
+      "@graphql-codegen/typescript-resolvers": "^1.15.1",
+      "@types/copy-webpack-plugin": "^6.0.0",
       "@types/cron": "^1.7.2",
       "@types/fs-extra": "^9.0.1",
-      "@types/koa": "^2.11.0",
+      "@types/koa": "^2.11.3",
       "@types/koa-router": "^7.4.1",
-      "@types/node": "^12.12.14",
-      "@types/react": "^16.9.15",
-      "@types/react-dom": "^16.9.4",
+      "@types/node": "^14.0.11",
+      "@types/react": "^16.9.35",
+      "@types/react-dom": "^16.9.8",
       "@types/react-redux": "^7.1.9",
-      "@types/react-router-dom": "^5.1.3",
-      "@types/workbox-sw": "^4.3.0",
-      "@typescript-eslint/eslint-plugin": "^2.10.0",
-      "@typescript-eslint/parser": "^2.10.0",
-      "@typescript-eslint/typescript-estree": "^2.10.0",
-      "babel-loader": "^8.0.6",
-      "copy-webpack-plugin": "^5.0.5",
-      "css-loader": "^3.2.1",
-      "eslint": "^6.7.2",
-      "eslint-config-airbnb": "^18.0.1",
+      "@types/react-router-dom": "^5.1.5",
+      "@types/terser-webpack-plugin": "^3.0.0",
+      "@types/webpack": "^4.41.17",
+      "@types/webpack-dev-server": "^3.11.0",
+      "@types/webpack-merge": "^4.1.5",
+      "@types/workbox-sw": "^4.3.1",
+      "@typescript-eslint/eslint-plugin": "^3.1.0",
+      "@typescript-eslint/parser": "^3.1.0",
+      "@typescript-eslint/typescript-estree": "^3.1.0",
+      "babel-loader": "^8.1.0",
+      "copy-webpack-plugin": "^6.0.2",
+      "css-loader": "^3.5.3",
+      "eslint": "^7.2.0",
+      "eslint-config-airbnb": "^18.1.0",
       "eslint-import-resolver-alias": "^1.1.2",
-      "eslint-import-resolver-webpack": "^0.12.0",
-      "eslint-plugin-graphql": "^3.1.0",
-      "eslint-plugin-import": "^2.18.2",
+      "eslint-import-resolver-webpack": "^0.12.1",
+      "eslint-plugin-graphql": "^4.0.0",
+      "eslint-plugin-import": "^2.20.2",
       "eslint-plugin-jsx-a11y": "^6.2.3",
-      "eslint-plugin-react": "^7.17.0",
-      "eslint-plugin-react-hooks": "^2.3.0",
-      "file-loader": "^5.0.2",
+      "eslint-plugin-react": "^7.20.0",
+      "eslint-plugin-react-hooks": "^4.0.4",
+      "file-loader": "^6.0.0",
       "hard-source-webpack-plugin": "^0.13.1",
-      "html-webpack-plugin": "^3.2.0",
+      "html-webpack-plugin": "^4.3.0",
       "jsdom": "^16.2.2",
-      "mini-css-extract-plugin": "^0.8.0",
-      "nodemon": "^2.0.1",
+      "mini-css-extract-plugin": "^0.9.0",
+      "nodemon": "^2.0.4",
       "npm-run-all": "^4.1.5",
-      "source-map-loader": "^0.2.4",
-      "terser-webpack-plugin": "^2.2.2",
-      "ts-loader": "^6.2.1",
-      "url-loader": "^3.0.0",
-      "webpack": "^4.41.2",
-      "webpack-bundle-analyzer": "^3.6.0",
-      "webpack-cli": "^3.3.10",
-      "webpack-dev-server": "^3.9.0",
+      "source-map-loader": "^1.0.0",
+      "terser-webpack-plugin": "^3.0.3",
+      "ts-loader": "^7.0.5",
+      "ts-node": "^8.10.2",
+      "typescript": "^3.9.5",
+      "url-loader": "^4.1.0",
+      "webpack": "^4.43.0",
+      "webpack-bundle-analyzer": "^3.8.0",
+      "webpack-cli": "^3.3.11",
+      "webpack-dev-server": "^3.11.0",
       "webpack-merge": "^4.2.2",
       "webpack-node-externals": "^1.7.2",
-      "workbox-sw": "^4.3.1",
-      "workbox-webpack-plugin": "^4.3.1"
+      "workbox-sw": "^5.1.3",
+      "workbox-webpack-plugin": "^5.1.3"
    },
    "dependencies": {
-      "@apollo/react-hooks": "^3.1.3",
-      "@babel/preset-react": "^7.7.4",
-      "@hot-loader/react-dom": "^16.11.0",
-      "@material-ui/core": "^4.10.0",
+      "@apollo/react-hooks": "^3.1.5",
+      "@babel/preset-react": "^7.10.1",
+      "@hot-loader/react-dom": "^16.13.0",
+      "@material-ui/core": "^4.10.1",
       "@material-ui/icons": "^4.9.1",
-      "@material-ui/lab": "^4.0.0-alpha.54",
-      "@nivo/line": "^0.61.1",
+      "@material-ui/lab": "^4.0.0-alpha.55",
+      "@nivo/line": "^0.62.0",
       "@reduxjs/toolkit": "^1.3.6",
       "@types/oauth2-server": "^3.0.12",
-      "apollo-cache-inmemory": "^1.6.3",
+      "apollo-cache-inmemory": "^1.6.6",
       "apollo-cache-persist": "^0.1.1",
-      "apollo-client": "^2.6.4",
-      "apollo-link": "^1.2.13",
+      "apollo-client": "^2.6.10",
+      "apollo-link": "^1.2.14",
       "apollo-link-context": "^1.0.20",
-      "apollo-link-error": "^1.1.12",
-      "apollo-link-ws": "^1.0.19",
-      "apollo-server-koa": "^2.9.13",
-      "apollo-upload-client": "^12.1.0",
-      "apollo-utilities": "^1.3.2",
+      "apollo-link-error": "^1.1.13",
+      "apollo-link-ws": "^1.0.20",
+      "apollo-server-koa": "^2.14.2",
+      "apollo-upload-client": "^13.0.0",
+      "apollo-utilities": "^1.3.4",
       "argon2": "^0.26.2",
       "axios": "^0.19.2",
       "clean-webpack-plugin": "^3.0.0",
       "cron": "^1.8.2",
-      "cross-env": "^6.0.3",
+      "cross-env": "^7.0.2",
       "debug": "^4.1.1",
-      "fs-extra": "^9.0.0",
-      "graphql": "^14.5.8",
+      "fs-extra": "^9.0.1",
+      "graphql": "^15.0.0",
       "graphql-iso-date": "^3.6.1",
       "graphql-subscriptions": "^1.1.0",
-      "koa": "^2.11.0",
+      "koa": "^2.12.0",
       "koa-bodyparser": "^4.3.0",
       "koa-router": "^9.0.1",
       "koa-static": "^5.0.0",
       "koa2-connect-history-api-fallback": "^0.1.2",
       "koa2-oauth-server": "^1.0.0",
       "oauth2-server": "^3.0.2",
-      "react": "^16.12.0",
-      "react-dom": "^16.12.0",
-      "react-hot-loader": "^4.12.18",
+      "react": "^16.13.1",
+      "react-dom": "^16.13.1",
+      "react-hot-loader": "^4.12.21",
       "react-redux": "^7.2.0",
-      "react-router-dom": "^5.1.2",
+      "react-router-dom": "^5.2.0",
       "redux": "^4.0.5",
       "redux-persist": "^6.0.0",
-      "sequelize": "^5.21.11",
+      "sequelize": "^5.21.12",
       "sqlite3": "^4.2.0",
       "subscriptions-transport-ws": "^0.9.16",
-      "ts-node": "^8.5.4",
-      "typescript": "^3.7.3",
-      "workbox-window": "^4.3.1",
+      "workbox-window": "^5.1.3",
       "yaml": "^1.10.0"
    }
 }
